:recycle: Separate file for internal functions
